+++ pkgs/intl4x/test/utils.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:intl4x/intl_test.dart';
+import 'package:intl4x/intl_testutils.dart';
 import 'package:test/test.dart';
 
 void testWithFormatting<T>(
